Conflicts:
	WebContent/Employee.jsp
	WebContent/left.jsp
	WebContent/newWorkerSalary_daiqixin.jsp
	build/classes/.gitignore
	build/classes/com/sms/action/EmployeeAction.class
	build/classes/com/sms/action/InitSalaryCompute.class
	build/classes/com/sms/action/LoginAction.class
	build/classes/com/sms/action/ManageTest.class
	build/classes/com/sms/action/NewEmployeeAction.class
	build/classes/com/sms/action/SalaryAction.class
	build/classes/com/sms/action/UserAction.class
	build/classes/com/sms/dao/IAddOrCutDao.class
	build/classes/com/sms/dao/IDepartmentChangeDao.class
	build/classes/com/sms/dao/IEmployeeDao.class
	build/classes/com/sms/dao/IFinalBonusDao.class
	build/classes/com/sms/dao/IFinalCheckDao.class
	build/classes/com/sms/dao/ILeaveSchoolPersonDao.class
	build/classes/com/sms/dao/IManageSalaryDao.class
	build/classes/com/sms/dao/IProfSalaryDao.class
	build/classes/com/sms/dao/IStartSalaryInfoDao.class
	build/classes/com/sms/dao/ITitleChangeDao.class
	build/classes/com/sms/dao/IUserDao.class
	build/classes/com/sms/dao/IWorkChangeDao.class
	build/classes/com/sms/dao/IWorkerSalaryDao.class
	build/classes/com/sms/dao/impl/AddOrCutDaoImp.class
	build/classes/com/sms/dao/impl/DepartmentChangeDaoImp.class
	build/classes/com/sms/dao/impl/EmployeeDaoImp.class
	build/classes/com/sms/dao/impl/FinalBonusDaoImp.class
	build/classes/com/sms/dao/impl/FinalCheckDaoImp.class
	build/classes/com/sms/dao/impl/LeaveSchoolPersonDaoImp.class
	build/classes/com/sms/dao/impl/ManageSalaryDaoImp.class
	build/classes/com/sms/dao/impl/ProfSalaryDaoImp.class
	build/classes/com/sms/dao/impl/StartSalaryInfoDaoImp.class
	build/classes/com/sms/dao/impl/TitleChangeDaoImp.class
	build/classes/com/sms/dao/impl/UserDaoImp.class
	build/classes/com/sms/dao/impl/WorkChangeDaoImp.class
	build/classes/com/sms/dao/impl/WorkerSalaryDaoImp.class
	build/classes/com/sms/entity/AddOrCut.class
	build/classes/com/sms/entity/DepartmentChange.class
	build/classes/com/sms/entity/Employee.class
	build/classes/com/sms/entity/ExperienceInfo.class
	build/classes/com/sms/entity/FinalBonus.class
	build/classes/com/sms/entity/FinalCheck.class
	build/classes/com/sms/entity/LeaveSchoolPerson.class
	build/classes/com/sms/entity/ManagePaylevelSalary.class
	build/classes/com/sms/entity/ManagePositionSalary.class
	build/classes/com/sms/entity/ManageSalaryChange.class
	build/classes/com/sms/entity/ProfessionalPaylevelSalary.class
	build/classes/com/sms/entity/ProfessionalPositionSalary.class
	build/classes/com/sms/entity/ProfessionalSalaryChange.class
	build/classes/com/sms/entity/Salary.class
	build/classes/com/sms/entity/StartSalaryInfo.class
	build/classes/com/sms/entity/TitleChange.class
	build/classes/com/sms/entity/User.class
	build/classes/com/sms/entity/WorkChange.class
	build/classes/com/sms/entity/WorkerPaylevelSalary.class
	build/classes/com/sms/entity/WorkerPositionSalary.class
	build/classes/com/sms/entity/WorkerSalaryChange.class
	build/classes/com/sms/exception/MyException.class
	build/classes/com/sms/filter/CharsetEncodingFilter.class
	build/classes/com/sms/interceptor/CheckLogin.class
	build/classes/com/sms/other/ComputeSalarySingleton.class
	build/classes/com/sms/security/Md5.class
	build/classes/com/sms/service/IAddOrCutManage.class
	build/classes/com/sms/service/IDepartmentChangeManage.class
	build/classes/com/sms/service/IEmployeeManage.class
	build/classes/com/sms/service/IFinalBonusManage.class
	build/classes/com/sms/service/IFinalCheckManage.class
	build/classes/com/sms/service/ILeaveSchoolPersonManage.class
	build/classes/com/sms/service/IManageSalaryManage.class
	build/classes/com/sms/service/IStartSalaryInfoManage.class
	build/classes/com/sms/service/ITitleChangeManage.class
	build/classes/com/sms/service/IUserManage.class
	build/classes/com/sms/service/IWorkChangeManage.class
	build/classes/com/sms/service/LevelSalaryChange.class
	build/classes/com/sms/service/impl/AddOrCutManageImp.class
	build/classes/com/sms/service/impl/DepartmentChangeManageImp.class
	build/classes/com/sms/service/impl/EmployeeManageImp.class
	build/classes/com/sms/service/impl/FinalBonusManageImp.class
	build/classes/com/sms/service/impl/FinalCheckManageImp.class
	build/classes/com/sms/service/impl/LeaveSchoolPersonManageImp.class
	build/classes/com/sms/service/impl/ManageSalaryManageImp.class
	build/classes/com/sms/service/impl/StartSalaryInfoManageImp.class
	build/classes/com/sms/service/impl/TitleChangeManageImp.class
	build/classes/com/sms/service/impl/UserManageImp.class
	build/classes/com/sms/service/impl/WorkChangeManageImp.class
	build/classes/com/sms/test/TestEmployeeDaoImp.class
	build/classes/config/struts-user.xml
	src/config/struts-user.xml

+++ src/com/sms/entity/StartSalaryInfo.java
@@ -25,8 +25,8 @@ public class StartSalaryInfo {
 	private String  salarySeries; //工资系列
 	private Integer positionLevel;//工资岗位级别
 	private Integer salaryLevel;//薪级
-	private Double positionSalary;//岗位工资
-	private Double levelSalary;//薪级工资
+	private Integer positionSalary;//岗位工资
+	private Integer levelSalary;//薪级工资
 	private String remarks;//备注
 	
 	@Id
@@ -135,18 +135,18 @@ public class StartSalaryInfo {
 	}
 	
 	@Column
-	public Double getPositionSalary() {
+	public Integer getPositionSalary() {
 		return positionSalary;
 	}
-	public void setPositionSalary(Double positionSalary) {
+	public void setPositionSalary(Integer positionSalary) {
 		this.positionSalary = positionSalary;
 	}
 	
 	@Column
-	public Double getLevelSalary() {
+	public Integer getLevelSalary() {
 		return levelSalary;
 	}
-	public void setLevelSalary(Double levelSalary) {
+	public void setLevelSalary(Integer levelSalary) {
 		this.levelSalary = levelSalary;
 	}
